Count sold quantity in 25kg equivalents, not raw lines
Summing line quantities counted a 50kg sack the same as a 25kg one, so
the figure disagreed with the Sold Quantity the Sales Incentives page
reports. Incentives normalise by weight: SUM(weight * quantity) / 25,
which is what SalesOrderIncentive already stores as its amount.

Left join products and guard the weight with COALESCE, mirroring the
?? 0 in ArInvoiceClass, so a line whose product is missing cannot drop
the whole order from the total. The result can be fractional, so format
it in the table and export it as a float.

+++ app/Exports/SalesReportExport.php
@@ -210,7 +210,7 @@ class SalesReportExport implements WithMultipleSheets
             $rows[] = [
                 (string) data_get($item, 'employee_name', ''),
                 (int) data_get($item, 'so_count', 0),
-                (int) data_get($item, 'sold_quantity', 0),
+                (float) data_get($item, 'sold_quantity', 0),
                 (float) data_get($item, 'so_total', 0),
                 (int) data_get($item, 'ar_count', 0),
                 (float) data_get($item, 'ar_total', 0),
